Fix broken setfocus in Extras
When exiting a screen created from the Extras group, the focus was set
to the ButtonGroup in the main screen, prohibiting any further action in
the popup.  The only way to continue was to either hit the back button
or press the down button.  Thiss seems to fix the problem.