Wait for server to process play update
If there's a player update request in the middle of happening, we
probably want to account for that as it may affect what's on the home
page. If it doesn't get processed in time, the home page will be stale
until navigating to another view and back again.

I opted for more immediate user feedback here, at the cost of potentially doing an
extra updateHomeRows call. We could only update the items when the
request completes, but that would affect users who run into the case
where the final update request doesn't change the homepage. During
testing, that was a common outcome when quitting a show during the
credits, or waiting for it to be done and quit on its own.