Remove line which causes Roku app Crash or Freeze with episodes
This line causes the Roku client to crash or freeze when the user picks an episode. On the latest git version of the code it freezes and on the stable release on the Roku marketplace, it crashes out or sometimes plays the media, but zoomed in significantly which makes it unwatchable.